src/Database: Use a Hash object to derive a key from the pathname
And store it (and a placeholder value - the modification time) in the
database.

A hash of the filename as the key may not be required or necessary here as
the pathname may be perfectly optimum as a key time in LMDB (we have to store
it anyway). Benchmarking in the glim test module should be added to compare the
key types.