src/Database: Revert to a hashed key and use a Tag object as the value
For the time being we revert to using a hashed key although no benchmarking
has yet been done.

We introduce the use of our simple Tag type here as the value to store
in the DB. THIS PATCH WON'T BUILD. For the build we still need to write
our own boost archiver for the Tag type -- GLIM uses boost serialisation.