src/test/suffix-array: Port across old code from another project
This is a bit-based suffix array toy I wrote several years ago.
Given an integer it computes a suffix array of the bits used
to comprise the original integer. Effectively this gives us a
longest common prefix (of substring bits) of the integer.

I could find no use for it! Perhaps it may come in handy here after
a little work?