src/Database: Refactor to allow different DB entry types
This patch attempts to fix the previous implementation where,
upon retrieval, a DB entry was assumed to be a tag yet it could
have been an image. Here we introduce an ID that we serialise
along with main entry data type so we can distinguish the items
later.

Note that the list_entries() function body has been temporarily
removed and will be reintroduced in a later patch.