src/Database: Introduce a generic visitor and prepare for further updates
We shall be using the traversal originally found in list_entries() for
updating the DB too -- so move this code into a visit() method and have
the visitors implement the code specific to the function necessary.

For this both Accessor and Entry now have, for the time being, public
visibility. This will hopefully revert later when I've figured out a
better solution that still hides these details.