src/Database: Port glim DB usage to lmdbxx
We must now serialise our own data into the lmdb::val.
Remove the body of the visit methods for the time being.
This also changes the interface; we make Entry a template type
and for the time being, remove the friend function operator<<.

Much code has been removed to simplify this port/refactor.